Fix supershell position bug
While dogfooding the latest sbt, I realized that sometimes debug lines
were still decreasing the padding of the supershell region. This would
cause stale tasks to persist at the bottom of the supershell area. I
relax the isDebugLine check to be safe. This means that sometimes the
supershell region will stay a bit larger than it needs to be with extra
blank lines, but that's better than having the stale tasks at the
bottom.